add live-reload capabilities
