+++ assets/css/style.css
@@ -1,96 +1,125 @@
 html {
-  background: #1f1f1f;
-  height: 100%;
-  font-family: #ebdbb2;
-  color: #ebdbb2;
+    background: #1f1f1f;
+    height: 100%;
+    width: 100%;
+    font-family: "Jetbrains Mono";
+    color: #ebdbb2;
 }
 
 .body {
-  max-width: 40rem;
-  font-family: "Jetbrains Mono";
-  margin: 0 auto;
-  padding-top: 1em;
-  text-align: justify;
+    max-width: 40rem;
+    margin: 0 auto;
+    padding-top: 1em;
+    padding: 1ch;
+    text-align: justify;
 }
 
 img {
-  max-width: 50rem;
+    max-width: 50rem;
 }
 
-h1,
-h2,
-h3,
-h4,
-h5 {
-  font-family: "Fira Code";
-  margin-bottom: 0.4rem;
-  color: #fabd2f;
+h1 {
+    font-family: "Fira Code";
+    color: #fb482f;
+}
+
+h2 {
+    font-family: "Fira Code";
+    color: #b8bb26;
+}
+
+h3 {
+    font-family: "Fira Code";
+    color: #d3869b;
 }
 
 a {
-  text-decoration: none;
-  color: #e2e2e2;
+    text-decoration: none;
+    color: #83a598;
 }
 
 blockquote {
-  background-color: #242424;
-  border-radius: 0.2em;
-  text-align: justify;
-  margin: 0.1rem;
-  padding: 0.4rem 1.2rem;
+    background-color: #242424;
+    border-radius: 0.2em;
+    text-align: justify;
+    margin: 0.1rem;
+    padding: 0.4rem 1.2rem 0.4rem 1.2rem;
 }
 
 ::-moz-selection {
-  background-color: #91919165;
+    background-color: #91919165;
 }
 
 ::selection {
-  background-color: #91919165;
+    background-color: #91919165;
 }
 
 .index {
-  text-align: center;
-  margin: 0;
-  padding: 20% 0;
-  font-size: 5ch;
+    max-width: 60rem;
+    margin: 2ch auto;
+    height: 100%;
+}
+
+.index ul {
+    margin: 0 auto;
 }
 
 .index a {
-  color: #ffe75e;
+    color: #83a598;
 }
 
 .index li {
-  display: inline-block;
-  list-style-type: none;
+    font-size: 4ch;
+    list-style-type: none;
 }
 
 .chroma,
 pre {
-  font-family: "Jetbrains Mono";
-  background-color: #272727;
-  max-width: 40rem;
-  padding: 1.4rem;
-  margin: 0.8rem 0;
-  border-radius: 0.2em;
+    font-family: "Jetbrains Mono";
+    background-color: #272727;
+    max-width: 60rem;
+    padding: 1.4rem;
+    margin: 0.8rem 0;
+    border-radius: 0.2em;
+}
+
+code {
+    max-width: 4em;
 }
 
 .kd,
 .kn {
-  color: #fb482f;
+    color: #fb482f;
 }
 
 .s {
-  color: #b8bb26;
+    color: #b8bb26;
 }
 
 .nf {
-  color: #83a598;
+    color: #83a598;
 }
 
 .kc {
-  color: #d3869b;
+    color: #d3869b;
 }
 
 .kt {
-  color: #fabd2f;
+    color: #fabd2f;
+}
+
+footer {
+    text-align: center;
+    position: absolute;
+    width: 98%;
+}
+
+footer img {
+    width: 5ch;
+}
+
+footer li {
+    bottom: 0;
+    display: inline-block;
+    list-style-type: none;
 }
