+++ assets/templates/index.tmpl
@@ -7,6 +7,7 @@
     <title>Index</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <script id="MathJax-script" async src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
+    {{.Script}}
     <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/style.css" />
 </head>
