+++ go.mod
@@ -4,7 +4,12 @@ go 1.14
 
 require (
 	github.com/alecthomas/chroma v0.8.1
+	github.com/fsnotify/fsnotify v1.4.9 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/jaschaephraim/lrserver v0.0.0-20171129202958-50d19f603f71
 	github.com/litao91/goldmark-mathjax v0.0.0-20200921072530-4c5dae64834a
+	github.com/smartystreets/goconvey v1.6.4 // indirect
 	github.com/yuin/goldmark v1.2.1
 	github.com/yuin/goldmark-highlighting v0.0.0-20200307114337-60d527fdb691
+	gopkg.in/fsnotify.v1 v1.4.7
 )
